+++ src/bet.h
@@ -67,7 +67,7 @@ static int advatoi (const char *s)
     /*
      * as long as the current character is a digit add to current number
      */
-    while ( isdigit(s[0]) )
+    while ( isdigit((int)s[0]) )
 	number = (number * 10) + (*s++ - '0');
 
     switch (UPPER(s[0]))
@@ -79,7 +79,7 @@ static int advatoi (const char *s)
     }
 
     /* if any digits follow k/m, add those too */
-    while ( isdigit(s[0]) && (multiplier > 1))
+    while ( isdigit((int)s[0]) && (multiplier > 1))
     {
 	/* the further we get to right, the less the digit 'worth' */
 	multiplier /= 10;
@@ -87,7 +87,7 @@ static int advatoi (const char *s)
     }
 
     /* return 0 if non-digit character was found, other than NULL */
-    if (s[0] != '\0' && !isdigit(s[0]))
+    if (s[0] != '\0' && !isdigit((int)s[0]))
 	return 0;
 
     /* anything left is likely extra digits (ie: 14k4443  -> 3 is extra) */
@@ -124,7 +124,7 @@ static int parsebet (const int currentbet, const char *s)
     if ( s[0] != '\0' )
     {
 	/* if first char is a digit, use advatoi */
-	if ( isdigit(s[0]) )
+      if ( isdigit((int) s[0]) )
 	    return (advatoi(s));
 	if ( s[0] == '+' )		/* add percent (default 25%) */
 	{
